Now generating a plan according to the weight each exercise has

+++ app/src/main/java/agdesigns/elevatefitness/data/exercise/Exercise.kt
@@ -17,7 +17,7 @@ data class Exercise(
     val image: Int = R.drawable.finish_workout,
     val description: String = "Description not available",
     val difficulty: ExerciseDifficulty = ExerciseDifficulty.INTERMEDIATE,
-    val probability: Float = 1f, // weight used when randomly selecting exercises
+    val probability: Double = 1.0, // weight used when randomly selecting exercises
     val variations: List<String> = emptyList()
 
 ) : Parcelable {
